+++ CHANGELOG.md
@@ -22,3 +22,4 @@
 - Button: Fixed Buttons in groups wrapped with popover from losing border radius
 - Button: Updated :active styles
 - Theme: Supports passing multiple theme fragments to the same provider
+- FormControl: Label margin-bottom: 0 for the last child for better composition